fix: name that makes sense for the class
